ingestor/requirements.txt
tenacity
SQLAlchemy
aiohttp
asyncpg
python-dotenv